Split `KeyLogFile` into its own module.
Make it clear that `KeyLog` doesn't depend on `std::{env,fs,io,path,sync}`.

Use `git diff main:rustls/src/keylog.rs rustls/src/keylogfile.rs` to
verify that none of the `KeyLogFile` code has changed at all.
